pyproject.toml
[project]
name = "cv-search-poc"
version = "0.1.0"
description = "Candidate search and planning tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.0.0",
    "openai>=1.42.0,<2",
    "faiss-cpu",
    "numpy>=1.26",
    "sentence-transformers",
    "streamlit",
    "python-pptx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = [{include = "cv_search", from = "src"}]

[tool.uv]
package = true
